Added models file
