+++ src/state/tripState.ts
@@ -1,27 +1,5 @@
 import { atom } from 'recoil';
-
-export interface Trip {
-  id: string;
-  name: string;
-  start_time: Date;
-  end_time: Date;
-  private: boolean;
-  notes: Note[];
-  routes: Route[];
-  item_lists: ItemList[];
-}
-
-interface Note {
-  id: string;
-}
-
-interface Route {
-  id: string;
-}
-
-interface ItemList {
-  id: string;
-}
+import { Trip } from '../models/Types';
 
 export const tripState = atom<Trip[]>({
   key: 'tripState',
